will ADD: more logic refinements to SummaryCard.tsx and other minor improvements and tests
