+++ frontend/src/pages/HomePage.tsx
@@ -6,11 +6,12 @@ import { RootState, AppDispatch } from "../../src/store/store";
 import { useEffect } from "react";
 //import { loadProducts } from "../../src/store/slices/localProductSlice";
 import { loadProducts } from "../../src/store/slices/productSlice";
-import Test from "../utils/Test";
 
 const HomePage: React.FC = () => {
   const dispatch: AppDispatch = useDispatch();
-  const { products, searchResults, loading, error } = useSelector((state: RootState) => state.products);
+  const { products, searchResults, loading, error } = useSelector(
+    (state: RootState) => state.products,
+  );
 
   useEffect(() => {
     dispatch(loadProducts());
@@ -20,7 +21,6 @@ const HomePage: React.FC = () => {
 
   return (
     <>
-      <Test />
       <Header showAdvertising={true} />
       <div className="mb-4 pr-10 pl-10">
         <AdvertisingCarousel />
