+++ frontend/src/pages/ShoppingCartPage.tsx
@@ -5,11 +5,14 @@ import RegisterForm from "../utils/RegisterForm";
 import SigninForm from "../utils/SigninForm";
 import { useNavigate } from "react-router-dom";
 import ProductInCart from "../components/ProductList/ProductInCart";
+import SummaryCard from "../utils/SummaryCard";
 
 export default function ShoppingCartPage() {
   const cartItems = useSelector((state: RootState) => state.cart.items);
   const navigate = useNavigate();
 
+
+  
   return (
     <div className="h-screen bg-red-50">
       <div className="flex h-[70vh] flex-col items-center justify-center gap-5 bg-red-100">
@@ -28,7 +31,10 @@ export default function ShoppingCartPage() {
             </span>
           </>
         ) : (
+          <div className="flex gap-3">
           <ProductInCart cartItems={cartItems} />
+          <SummaryCard />
+          </div>
         )}
       </div>
       <div className="flex items-center justify-center gap-20 bg-gray-200">
