+++ frontend/src/utils/Button.tsx
@@ -18,7 +18,7 @@ const Button: React.FC<ButtonProps> = ({
   return (
     <button
       onClick={onClick}
-      className={`bg-primary hover:bg-primaryHover w-56 cursor-pointer rounded-full border border-black px-4 py-2.5 text-lg text-white transition ${className}`}
+      className={`bg-primary hover:bg-primaryHover w-56 cursor-pointer rounded-full border border-black px-4 py-2.5 text-lg text-red-800 transition ${className}`}
       disabled={disabled}
       type={type}
     >
